+++ src/sentence/output_format.cpp
@@ -55,7 +55,8 @@ void output_format_conllu::write_sentence(const sentence& s, ostream& os) {
           i == s.multiword_tokens[multiword_token].id_first) {
         os << s.multiword_tokens[multiword_token].id_first << '-'
            << s.multiword_tokens[multiword_token].id_last << '\t';
-        write_with_spaces(os, s.multiword_tokens[multiword_token].form) << "\t_\t_\t_\t_\t_\t_\t_\t"
+        write_with_spaces(os, s.multiword_tokens[multiword_token].form) << "\t_\t_\t_\t"
+           << underscore_on_empty(s.multiword_tokens[multiword_token].feats) << "\t_\t_\t_\t"
            << underscore_on_empty(s.multiword_tokens[multiword_token].misc) << '\n';
         multiword_token++;
       }