+++ CHANGES.md
@@ -1,5 +1,8 @@
 Version 1.3.2-dev
 -----------------
+- Allow Typo=Yes annotation in FEATS of a multi-word token.
+  - The `multiword_token` now has a new `feats` field.
+  - The `multiword_token.feats` is utilized both in CoNLL-U input and output format.
 
 
 Version 1.3.1 [15 Nov 23]